feat(#15): Add popup menu for vault intention and context menu action
