--- src/main/kotlin/de/timo_reymann/ansible_vault_integration/action/commandline/AnsibleCommandLineTransformer.kt
@@ -1,9 +1,0 @@
-package de.timo_reymann.ansible_vault_integration.action.commandline
-
-import com.intellij.execution.configurations.GeneralCommandLine
-import com.intellij.openapi.project.Project
-
-interface AnsibleCommandLineTransformer {
-    fun transformFileName(input: String?): String?
-    fun transformCommandLine(project: Project, commandLine: GeneralCommandLine): GeneralCommandLine? = commandLine
-}
