--- src/main/kotlin/de/timo_reymann/ansible_vault_integration/action/commandline/NoOpAnsibleCommandLineTransformer.kt
@@ -1,5 +1,0 @@
-package de.timo_reymann.ansible_vault_integration.action.commandline
-
-class NoOpAnsibleCommandLineTransformer : AnsibleCommandLineTransformer {
-    override fun transformFileName(input: String?): String? = input
-}
