--- src/main/kotlin/de/timo_reymann/ansible_vault_integration/action/execution/AnsibleVaultWrapperCallFailedException.kt
@@ -1,3 +1,0 @@
-package de.timo_reymann.ansible_vault_integration.action.execution
-
-class AnsibleVaultWrapperCallFailedException(message: String?) : Exception(message)
