--- src/main/kotlin/de/timo_reymann/ansible_vault_integration/action/runnable/AnsibleVaultRunnable.kt
@@ -1,20 +1,0 @@
-package de.timo_reymann.ansible_vault_integration.action.runnable
-
-/**
- * Abstraction for ansible vault task
- */
-interface AnsibleVaultRunnable {
-    /***
-     * Execute task using ansible vault
-     * @throws Exception Exceptions during execution
-     */
-    @Throws(Exception::class)
-    fun run()
-
-    /**
-     * Get message for success
-     *
-     * @return String
-     */
-    val successMessage: String
-}
