+++ src/main/kotlin/de/timo_reymann/ansible_vault_integration/commandline/NoOpAnsibleCommandLineTransformer.kt
@@ -1,0 +1,5 @@
+package de.timo_reymann.ansible_vault_integration.commandline
+
+class NoOpAnsibleCommandLineTransformer : AnsibleCommandLineTransformer {
+    override fun transformFileName(input: String?): String? = input
+}
