+++ src/main/kotlin/de/timo_reymann/ansible_vault_integration/execution/AnsibleVaultWrapperCallFailedException.kt
@@ -1,0 +1,3 @@
+package de.timo_reymann.ansible_vault_integration.execution
+
+class AnsibleVaultWrapperCallFailedException(message: String?) : Exception(message)
