+++ src/test/kotlin/de/timo_reymann/ansible_vault_integration/config/AnsibleConfigurationFileParserTest.kt
@@ -1,8 +1,5 @@
 package de.timo_reymann.ansible_vault_integration.config
 
-import de.timo_reymann.ansible_vault_integration.action.config.AnsibleConfigurationFileParser
-import de.timo_reymann.ansible_vault_integration.action.config.VaultIdentity
-import org.junit.Assert.assertArrayEquals
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import java.io.File
@@ -16,23 +13,23 @@ class AnsibleConfigurationFileParserTest {
 
     @Test
     fun `Verify empty file does not crash the parser`() {
-        AnsibleConfigurationFileParser(getFile("ansible.cfg"))
+        AnsibleConfigurationFileParser(getFile("ansible.cfg"), 0)
     }
 
     @Test
     fun `Verify stock config file does not crash the parser`() {
-        AnsibleConfigurationFileParser(getFile("ansible-stock.cfg"))
+        AnsibleConfigurationFileParser(getFile("ansible-stock.cfg"), 0)
     }
 
     @Test
     fun `Validate checksum calculation works`() {
         assertEquals(
             "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
-            AnsibleConfigurationFileParser(getFile("ansible.cfg")).calculateChecksum()
+            AnsibleConfigurationFileParser(getFile("ansible.cfg"), 0).calculateChecksum()
         )
         assertEquals(
             "843071daf968ddc5452bc54780a651a527792271be400d1e3a31e33dc11f1bdd",
-            AnsibleConfigurationFileParser(getFile("ansible-stock.cfg")).calculateChecksum()
+            AnsibleConfigurationFileParser(getFile("ansible-stock.cfg"), 0).calculateChecksum()
         )
     }
 
@@ -44,7 +41,7 @@ class AnsibleConfigurationFileParserTest {
             listOf("[defaults]")
         )
 
-        val parser = AnsibleConfigurationFileParser(tempFile.toFile())
+        val parser = AnsibleConfigurationFileParser(tempFile.toFile(), 0)
         val config = parser.parse()
 
         assertFalse("No change should be stated when there is none") { parser.hasChangedOnDisk(config) }
@@ -61,16 +58,40 @@ class AnsibleConfigurationFileParserTest {
 
     @Test
     fun `Verify vault identity list parsing doesnt crash on empty list`() {
-        val parser = AnsibleConfigurationFileParser(getFile("ansible.cfg"))
+        val parser = AnsibleConfigurationFileParser(getFile("ansible.cfg"), 0)
         parser.parseVaultIdentityList()
     }
 
     @Test
     fun `Verify vault identities are getting parsed correctly`() {
-        val parser = AnsibleConfigurationFileParser(getFile("ansible-vault-identities.cfg"))
-        assertEquals(listOf(
-            VaultIdentity("my_first_vault","~/ansible/passwords/my_first_vault"),
-                    VaultIdentity("my_second_vault","~/ansible/passwords/my_second_vault")
-        ), parser.parseVaultIdentityList())
+        val parser = AnsibleConfigurationFileParser(getFile("ansible-vault-identities.cfg"), 0)
+        assertEquals(
+            listOf(
+                VaultIdentity("my_first_vault", "~/ansible/passwords/my_first_vault"),
+                VaultIdentity("my_second_vault", "~/ansible/passwords/my_second_vault")
+            ), parser.parseVaultIdentityList()
+        )
+    }
+
+    @Test
+    fun `Verify merge works with one empty`() {
+        val merged = mergeConfigs(
+            listOf(
+                AnsibleConfigurationFile(listOf(VaultIdentity("test", "test")), "123", 1),
+                AnsibleConfigurationFile(null, "123", 0)
+            )
+        )
+        assertEquals(AnsibleConfigurationFile(listOf(VaultIdentity("test", "test")),null,null), merged)
+    }
+
+    @Test
+    fun `Verify merge works with priority`() {
+        val merged = mergeConfigs(
+            listOf(
+                AnsibleConfigurationFile(listOf(VaultIdentity("test", "test")), "123", 1),
+                AnsibleConfigurationFile(listOf(VaultIdentity("test0", "test0")), "123", 0)
+            )
+        )
+        assertEquals(AnsibleConfigurationFile(listOf(VaultIdentity("test", "test")),null,null), merged)
     }
 }
